The UI Works now.
So. The tab buttons toggle their respective pages when clicked. They also hide themselves and the other buttons, if they share the same space as the to be displayed page. 
to reset everything back to normal after making a selection, the back button has a reset group script. That one has all the TabButtons and their pages subscribe to it when they appear, and it remembers if they are active or inactive at the time. when its reset function is called, it puts all the subscribers back in the state that they entered the list in