Continued on PhasechangeTile. Almost done
they have the stressFreinds function now and call it yutomatically. RN the job is to find a way for the mapBuilder to tell them about their StressFieldStrength and so on, for that those variables probably neet to be movet to the main MapTile Class so the Builder can acess them. Should have just made all the tiles the same class honestly but welp.