Solved Red Errors. "Solved" Giant Mystery
I changed the matrix tile to acess not the globalStress vector directly from the eventManager but instead added an acessor function for it. now it works as it should. before, whatever vector was in eventmanager was being completely ignored and the tile just used the right pointing vector, despite having NO REFERENCE TO IT WHATSOEVER. it also didnt complain when the vector in eventmanager that it was supposed to use for its setup didnt exist (because it was using another one anyways)