Having fun with stresses still
added the code for the itnernal stress inside the particle. first the problem was the range offset (automatic cutoff when stress gets too weak works now btw), the offset didnt work right when the thing was inside the particle, so I added a check for the rangemod to see who called the function basically.

Now the problem is, that  the stressing out happens when the first tile is placed, so the other particle tiles arent there yet to see it happen. ugh! but i found a solution. Just create a coroutine that waits for the next frame to call the method. genious!