Still fun with stresses. Now in their own dictionary
I realized that the way it was, the stresses would be gone if you replaced an area with a particle after the phase change tile was placed. 
So I wrote the whole thing to be in its own ledger, indipendantly from the tile. had some hickups, the build function of the map builder now only places however many tiles it should, the thing works so far, but for some reason the game crashes when a partivcel tile is removed. which can only be done by map reset. I suspect the removeStress function to do wierd things when it doesnt find the sress or something