Question 1: why does my phase change particle not get deleted even tho it should behave like any other maptile child would? also, why does the game crash when it does get deleted by the map builder? (its what happens onDisable())
So much FUN! we have narrowed the crashing down to that when we try to remove the entry in the list of the position, something else is probably also acessing the stressdict right now (tho I can really only be the same function that is running atm or some read only stuff)

Right now trying to restructure the thing a bit, by making the stressoutFriends() function not enter or remove anything, It just creates a dictionary with the positions and the respective stress vectors to be added/removed.

that list is passed to the tilesdict which then adds/removes the entries from that list in its own function. No idea if that changes anything but it only took 30 mins to type up so might aswell try