Why did I put in a mechanic called stress?
could have seen that coming. anyways, I have figuted out that the crashing had something to do with either iterating over the dictionary whilst changing entries in it OR that it for some reason just really doesnt like when you edit an entry in a list that is part of a dictionary. dk why, thats like its functinonality but oh well. Its not crashing anymore, since I made a copy of the dictionary to iterate over so that one doesnt change during the process itslef, and also instead of editiing the list directly I amnow copying the whole list, editing the copy, removing the corresponding entry in the dictionary and then creating a new entry in its place, with the modified list as the value.
next problem however is, that its not removing all of the stresses. It does some and not others. I thought it was due to the vektors not being found, but turns out it just never tries to. the function doesnt get called half as often as it should. which is wierd. Im gonna next try the approach of creating the list with all the stresses, passing that over to the dict and letting it edit himself. that should prevent the thing stopping somewhere in the middle. unless it crashes again, that is