Zu früh gefreut.
Crashes when map gets rebuilt, because after the deleting of the stress, the tilesDict is asked for a tile so that can be updated as to its stress but the tileDict is changing while that happens because the map is being rebuilt. But I fixed it like a genius. Instead of updating the tile directly, which needs access to the tilesDict, the changed position is written into a private list, which is processed every update, if it has any entries. all the tiles at the positions in the list get uodated their stresses and then the list is cleared