More levels, added large hex map and button shifting
the play button now also acts as the refresh button. By just having the reset function in crack listen to the play button instead of the reset one. and the tile clearing in the map builder is now a public function and gets called as part of the reset that the crack is doing, instead of having its function listen to the reset/play button